--- services/painter/src/render/pil.py
@@ -1,16 +1,0 @@
-from PIL import Image, ImageDraw
-
-
-def render_rects(rects, canvas_size):
-    canvas = Image.new('RGB', canvas_size, "white")
-
-    draw = ImageDraw.Draw(canvas)
-
-    for rect in rects:
-        rectangle = (rect.x, rect.y,
-                     rect.x + rect.width,
-                     rect.y + rect.height)
-
-        draw.rectangle(rectangle)
-
-    canvas.show()
