+++ services/painter/src/util.py
@@ -1,0 +1,3 @@
+# https://stackoverflow.com/questions/952914/how-to-make-a-flat-list-out-of-a-list-of-lists
+def flatten(t):
+    return [item for sublist in t for item in sublist]
